requirements.txt
binance-futures-connector>=3.5.1
python-dotenv